+++ encoding/thrift/doc.go
@@ -34,11 +34,12 @@
 //
 // In addition to generating code for types specified in your Thrift file,
 // this will generate two packages for each service in the file: a client
-// package and a server package.
+// package, a server package, and a test package.
 //
 // 	myservice
 // 	   |- myserviceclient
 // 	   |- myserviceserver
+// 	   |- myservicetest
 //
 // The client package allows sending requests through a YARPC dispatcher.
 //
@@ -49,4 +50,10 @@
 //
 // 	handler := myHandler{}
 // 	dispatcher.Register(myserviceserver.New(handler))
+//
+// The test package provides a gomock-compatible mock client for the service.
+//
+// 	mockCtrl := gomock.NewController(t)
+// 	client := myservicetest.NewMockClient(mockCtrl)
+// 	client.EXPECT().Hello(request).Return(response, nil)
 package thrift