+++ encoding/thrift/thriftrw-plugin-yarpc/main.go
@@ -43,12 +43,17 @@ var (
 	_onewayHandlerWrapper = flag.String("oneway-handler-wrapper",
 		"go.uber.org/yarpc/encoding/thrift.OnewayHandler",
 		"Function used to wrap generic Thrift oneway function handlers into YARPC handlers")
+	_noGomock = flag.Bool("no-gomock", false,
+		"Don't generate gomock mocks for service clients")
 )
 
 type g struct{}
 
 func (g) Generate(req *api.GenerateServiceRequest) (*api.GenerateServiceResponse, error) {
 	generators := []genFunc{clientGenerator, serverGenerator}
+	if !*_noGomock {
+		generators = append(generators, gomockGenerator)
+	}
 
 	unaryWrapperImport, unaryWrapperFunc := splitFunctionPath(*_unaryHandlerWrapper)
 	onewayWrapperImport, onewayWrapperFunc := splitFunctionPath(*_onewayHandlerWrapper)