+++ encoding/thrift/thriftrw-plugin-yarpc/roundtrip_test.go
@@ -31,6 +31,8 @@ import (
 	"go.uber.org/yarpc"
 	"go.uber.org/yarpc/api/transport"
 	"go.uber.org/yarpc/encoding/thrift/thriftrw-plugin-yarpc/internal/tests/atomic"
+	"go.uber.org/yarpc/encoding/thrift/thriftrw-plugin-yarpc/internal/tests/atomic/readonlystoreclient"
+	"go.uber.org/yarpc/encoding/thrift/thriftrw-plugin-yarpc/internal/tests/atomic/readonlystoreserver"
 	"go.uber.org/yarpc/encoding/thrift/thriftrw-plugin-yarpc/internal/tests/atomic/storeclient"
 	"go.uber.org/yarpc/encoding/thrift/thriftrw-plugin-yarpc/internal/tests/atomic/storeserver"
 	"go.uber.org/yarpc/encoding/thrift/thriftrw-plugin-yarpc/internal/tests/common/baseserviceclient"
@@ -39,6 +41,8 @@ import (
 	"go.uber.org/yarpc/encoding/thrift/thriftrw-plugin-yarpc/internal/tests/common/emptyserviceserver"
 	"go.uber.org/yarpc/encoding/thrift/thriftrw-plugin-yarpc/internal/tests/common/extendemptyclient"
 	"go.uber.org/yarpc/encoding/thrift/thriftrw-plugin-yarpc/internal/tests/common/extendemptyserver"
+	"go.uber.org/yarpc/encoding/thrift/thriftrw-plugin-yarpc/internal/tests/common/extendonlyclient"
+	"go.uber.org/yarpc/encoding/thrift/thriftrw-plugin-yarpc/internal/tests/common/extendonlyserver"
 	"go.uber.org/yarpc/transport/http"
 
 	"github.com/stretchr/testify/assert"
@@ -79,9 +83,23 @@ func TestRoundTrip(t *testing.T) {
 			wantResult:    true,
 		},
 		{
+			desc:          "extend only: healthy",
+			procedures:    extendonlyserver.New(&storeHandler{healthy: true}),
+			newClientFunc: extendonlyclient.New,
+			method:        "Healthy",
+			wantResult:    true,
+		},
+		{
 			desc:          "store: healthy",
 			procedures:    storeserver.New(&storeHandler{healthy: true}),
 			newClientFunc: storeclient.New,
+			method:        "Healthy",
+			wantResult:    true,
+		},
+		{
+			desc:          "store: healthy with base client",
+			procedures:    storeserver.New(&storeHandler{healthy: true}),
+			newClientFunc: baseserviceclient.New,
 			method:        "Healthy",
 			wantResult:    true,
 		},
@@ -135,12 +153,48 @@ func TestRoundTrip(t *testing.T) {
 			},
 		},
 		{
-			desc:          "store: integer",
+			desc:          "store: integer with readonly client",
 			procedures:    storeserver.New(&storeHandler{integer: 42}),
+			newClientFunc: readonlystoreclient.New,
+			method:        "Integer",
+			methodArgs:    []interface{}{ptr.String("foo")},
+			wantResult:    int64(42),
+		},
+		{
+			desc: "readonly store: integer error with rw client",
+			procedures: readonlystoreserver.New(&storeHandler{
+				failWith: &atomic.KeyDoesNotExist{Key: ptr.String("foo")},
+			}),
+			newClientFunc: storeclient.New,
+			method:        "Integer",
+			methodArgs:    []interface{}{ptr.String("foo")},
+			wantError:     &atomic.KeyDoesNotExist{Key: ptr.String("foo")},
+		},
+		{
+			desc:          "readonly store: integer with readonly client",
+			procedures:    readonlystoreserver.New(&storeHandler{integer: 42}),
+			newClientFunc: readonlystoreclient.New,
+			method:        "Integer",
+			methodArgs:    []interface{}{ptr.String("foo")},
+			wantResult:    int64(42),
+		},
+		{
+			desc:          "readonly store: integer with rw client",
+			procedures:    readonlystoreserver.New(&storeHandler{integer: 42}),
 			newClientFunc: storeclient.New,
 			method:        "Integer",
 			methodArgs:    []interface{}{ptr.String("foo")},
 			wantResult:    int64(42),
+		},
+		{
+			desc: "readonly store: integer failure with rw client",
+			procedures: readonlystoreserver.New(&storeHandler{
+				failWith: &atomic.KeyDoesNotExist{Key: ptr.String("foo")},
+			}),
+			newClientFunc: storeclient.New,
+			method:        "Integer",
+			methodArgs:    []interface{}{ptr.String("foo")},
+			wantError:     &atomic.KeyDoesNotExist{Key: ptr.String("foo")},
 		},
 		{
 			desc: "store: integer failure",